Working on v1 score.
