+++ torchbenchmark/score/configs/v1/config-v1.env
@@ -1,0 +1,23 @@
+# This is the file that tunes the benchmark configurations
+
+# Config version
+CONFIG_VER=v1
+# Python and CUDA Version
+PYTHON_VERSION=3.7
+CUDA_VERSION=cu102
+# Number of iterations
+NUM_ITER=1
+# Number of rounds (minimum times to run each benchmark)
+NUM_ROUNDS=20
+# The CPU cores that run the benchmark
+# On the g4dn.metal machine, limit the CPU cores to single NUMA node
+CORE_LIST=24-47
+# The GPU devices that run the benchmark
+GPU_LIST=0
+# The <memory, graphics> clock frequency setting of GPU
+GPU_FREQUENCY=5001,900
+
+# TODO: Benchmark filter
+BENCHMARK_FILTER="(not slomo)"
+# Config file
+CONFIG_FILE=config-v1.yaml
